与shine联调代码修改

+++ rcdc-terminal-module-def/src/main/java/com/ruijie/rcos/rcdc/terminal/module/def/api/dto/CbbTerminalDetectDTO.java
@@ -41,7 +41,7 @@ public class CbbTerminalDetectDTO {
     /**
      * 网络时延
      */
-    private Double delay;
+    private Integer delay;
 
     /**
      * 检测状态
@@ -152,12 +152,12 @@ public class CbbTerminalDetectDTO {
     }
 
 
-    public Double getDelay() {
+    public Integer getDelay() {
         return delay;
     }
 
 
-    public void setDelay(Double delay) {
+    public void setDelay(Integer delay) {
         this.delay = delay;
     }
